+++ src/literalizer/languages/common_lisp.py
@@ -316,6 +316,13 @@ class CommonLisp(metaclass=LanguageCls):
 
     heterogeneous_strategies = HeterogeneousStrategies
 
+    class VersionFormats(enum.Enum):
+        """Version options for Common Lisp."""
+
+        ANSI = enum.auto()
+
+    version_formats = VersionFormats
+
     identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
         IdentifierCase.KEBAB,
     )
@@ -375,6 +382,7 @@ class CommonLisp(metaclass=LanguageCls):
     heterogeneous_strategy: HeterogeneousStrategies = (
         HeterogeneousStrategies.ERROR
     )
+    language_version: VersionFormats = VersionFormats.ANSI
     indent: str = "    "
 
     null_literal: ClassVar[str] = "nil"