+++ src/literalizer/languages/rust.py
@@ -1201,6 +1201,13 @@ class Rust(metaclass=LanguageCls):
 
     heterogeneous_strategies = HeterogeneousStrategies
 
+    class VersionFormats(enum.Enum):
+        """Version options for Rust."""
+
+        EDITION_2021 = enum.auto()
+
+    version_formats = VersionFormats
+
     identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
         IdentifierCase.SNAKE,
         IdentifierCase.PASCAL,
@@ -1267,6 +1274,7 @@ class Rust(metaclass=LanguageCls):
         HeterogeneousStrategies.ERROR
     )
     heterogeneous_value_enum_name: str = "Value"
+    language_version: VersionFormats = VersionFormats.EDITION_2021
     indent: str = "    "
 
     null_literal: ClassVar[str] = "None::<()>"