+++ src/literalizer/languages/sml.py
@@ -524,6 +524,13 @@ class Sml(metaclass=LanguageCls):
 
     heterogeneous_strategies = HeterogeneousStrategies
 
+    class VersionFormats(enum.Enum):
+        """Version options for SML."""
+
+        SML_97 = enum.auto()
+
+    version_formats = VersionFormats
+
     identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
         IdentifierCase.SNAKE,
         IdentifierCase.CAMEL,
@@ -583,6 +590,7 @@ class Sml(metaclass=LanguageCls):
     heterogeneous_strategy: HeterogeneousStrategies = (
         HeterogeneousStrategies.ERROR
     )
+    language_version: VersionFormats = VersionFormats.SML_97
     indent: str = "    "
     type_name: str = "val_t"
     constructor_prefix: str = "S"